extract VersionUtils class.
