+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/source/HttpSourceConnector.java
@@ -1,5 +1,6 @@
 package io.github.clescot.kafka.connect.http.source;
 
+import io.github.clescot.kafka.connect.http.VersionUtils;
 import io.github.clescot.kafka.connect.http.core.queue.QueueFactory;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
@@ -44,6 +45,6 @@ public class HttpSourceConnector extends SourceConnector {
 
     @Override
     public String version() {
-        return QueueFactory.VersionUtil.version(this.getClass());
+        return VersionUtils.version(this.getClass());
     }
 }
