+++ kafka-connect-http-core/src/main/java/io/github/clescot/kafka/connect/http/core/HttpExchange.java
@@ -8,7 +8,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class HttpExchange implements Serializable {
 
     public static final long serialVersionUID = 1L;
-    public static final String SCHEMA_ID = "https://raw.githubusercontent.com/clescot/kafka-connect-http/master/kafka-connect-http-core/src/main/resources/http-exchange.json";
+    public static final String BASE_SCHEMA_ID = "https://raw.githubusercontent.com/clescot/kafka-connect-http/master/kafka-connect-http-core/src/main/resources/schemas/";
+    public static final String SCHEMA_ID = BASE_SCHEMA_ID+"http-exchange.json";
     public static final String SCHEMA_AS_STRING = "{\n" +
             "  \"$id\": \"" + SCHEMA_ID + "\",\n" +
             "  \"$schema\": \"http://json-schema.org/draft/2019-09/schema\",\n" +
